+++ packages/common/src/relay/Relay.ts
@@ -6,23 +6,33 @@ export default abstract class Relay {
 
   protected abstract service: string
 
+  abstract notificationHandler(notification: any): void
+
   constructor(public session: BaseSession) {
 
   }
 
-  async registerContexts(contexts: string | string[], handler: Function) {
-    if (typeof contexts === 'string') {
-      contexts = [contexts]
-    }
-
+  async onReceive(contexts: string[], handler: Function) {
     const success = await Receive(this.session, contexts)
     if (success) {
-      contexts.forEach(context => register(this.session.relayProtocol, handler, this._ctxUniqueId(context)))
+      contexts.forEach(context => register(this.session.relayProtocol, handler, this._ctxReceiveUniqueId(context)))
     }
     return success
   }
 
-  protected _ctxUniqueId(context: string): string {
-    return `${this.service}.ctx.${context}`
+  async onStateChange(contexts: string[], handler: Function) {
+    const success = await Receive(this.session, contexts)
+    if (success) {
+      contexts.forEach(context => register(this.session.relayProtocol, handler, this._ctxStateUniqueId(context)))
+    }
+    return success
+  }
+
+  protected _ctxReceiveUniqueId(context: string): string {
+    return `${this.service}.ctxReceive.${context}`
+  }
+
+  protected _ctxStateUniqueId(context: string): string {
+    return `${this.service}.ctxState.${context}`
   }
 }