+++ packages/common/src/util/interfaces.ts
@@ -288,8 +288,38 @@ export interface IRelayConsumerParams {
   token: string
   contexts?: string[]
   onIncomingCall?: Function
+  onIncomingMessage?: Function
+  onMessageStateChange?: Function
   onTask?: Function
   setup?: Function
   ready?: Function
   teardown?: Function
+}
+
+export interface IMessage {
+  id: string
+  state: string
+  context: string
+  from: string
+  to: string
+  direction: string
+  tags: string[]
+  body: string
+  media: string[]
+  segments: number
+}
+
+
+export interface IMessageOptions {
+  message_id: string
+  message_state: string
+  context: string
+  from_number: string
+  to_number: string
+  direction: string
+  tags: string[]
+  body: string
+  media: string[]
+  segments: number
+  reason?: string
 }