+++ packages/node/src/relay/index.ts
@@ -1,6 +1,7 @@
 import BaseSession from '../../../common/src/BaseSession'
 import Calling from '../../../common/src/relay/calling/Calling'
 import Tasking from '../../../common/src/relay/tasking/Tasking'
+import Messaging from '../../../common/src/relay/messaging/Messaging'
 import { ISignalWireOptions } from '../../../common/src/util/interfaces'
 import logger from '../../../common/src/util/logger'
 
@@ -8,6 +9,7 @@ export default class RelayClient extends BaseSession {
 
   private _calling: Calling = null
   private _tasking: Tasking = null
+  private _messaging: Messaging = null
 
   constructor(public options: ISignalWireOptions) {
     super(options)
@@ -26,6 +28,13 @@ export default class RelayClient extends BaseSession {
       this._tasking = new Tasking(this)
     }
     return this._tasking
+  }
+
+  get messaging(): Messaging {
+    if (!this._messaging) {
+      this._messaging = new Messaging(this)
+    }
+    return this._messaging
   }
 
   private _handleSignals(): void {