+++ packages/node/tests/relay/calling.test.ts
@@ -137,23 +137,23 @@ describe('Calling', () => {
     })
   })
 
-  describe('.registerContexts()', () => {
+  describe('.onReceive()', () => {
 
     it('should register the inbound listener', async done => {
       Connection.mockSend.mockClear()
-      await session.calling.registerContexts('context', jest.fn())
+      await session.calling.onReceive(['context'], jest.fn())
 
       expect(Connection.mockSend).toHaveBeenCalledTimes(1)
       const { method } = Connection.mockSend.mock.calls[0][0].request.params
       expect(method).toEqual('signalwire.receive')
-      expect(isQueued(session.relayProtocol, `calling.ctx.context`)).toEqual(true)
+      expect(isQueued(session.relayProtocol, 'calling.ctxReceive.context')).toEqual(true)
 
       done()
     })
 
     it('should handle the calling.call.receive notification', async done => {
       const fnMock = jest.fn()
-      await session.calling.registerContexts('context', fnMock)
+      await session.calling.onReceive(['context'], fnMock)
       const msg = JSON.parse('{"id":"a16dae67-212b-4391-bb78-a0b9e45310b9","jsonrpc":"2.0","method":"blade.broadcast","params":{"broadcaster_nodeid":"9811eb32-1234-1234-1234-ab56fa3b83c9","channel":"notifications","event":"queuing.relay.events","params":{"event_channel":"signalwire_service_random_uuid","event_type":"calling.call.receive","params":{"call_id":"849982ab-1234-5678-1234-311534fa20d6","call_state":"created","context":"context","device":{"params":{"from_number":"+12222222222","to_number":"+12222222223"},"type":"phone"},"node_id":"9811eb32-1234-5678-XXXX-ab56fa3b83c9"},"project_id":"project","space_id":"space","timestamp":1149889452.302629},"protocol":"signalwire_service_random_uuid"}}')
       trigger(SwEvent.SocketMessage, msg, session.uuid)
       expect(fnMock).toHaveBeenCalledTimes(1)